requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.60.1
django-cache-url==3.4.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
environs==10.3.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.20.2
oauthlib==3.2.2
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0
